requirements.txt
# Schedule Agent Dependencies
# Core
pyyaml>=6.0
pydantic>=2.0
python-dateutil>=2.8

# LLM
openai>=1.0
httpx>=0.24
mcp>=1.0

# CLI
prompt_toolkit>=3.0
rich>=14.0

# Testing
pytest>=7.0
pytest-asyncio>=0.21

# Development
black>=23.0
isort>=5.12
